box_firmware: the great rename
