+++ hardware/firmware/box_rp2040/src/display/colour_schemes.h
