+++ hardware/firmware/box_rp2040/src/main.c
@@ -3,10 +3,10 @@
 #include "pico/stdlib.h"
 #include "pico/st7789.h"
 
-#include "usb.h"
-#include "display.h"
-#include "io.h"
-#include "network_switch_status_reader.h"
+#include "usb/usb.h"
+#include "display/display.h"
+#include "io/serial.h"
+#include "network_switch/network_switch_status_reader.h"
 
 void init(void) {
     display_init();
