+++ hardware/firmware/box_rp2040/src/network_switch/network_switch_status_reader.h
